Rename messages module to protobuf
